fix: clean cfa list + move script
