+++ server/src/jobs/offrePartenaire/isCompanyInBlockedCfaList.ts
@@ -42,7 +42,6 @@ export const cfaCompanyList = [
   "ACQUIFORMATIONS",
   "ACT'IN CAMPUS PRO",
   "ACTION FORMATION 81",
-  "ACTION FORMATION",
   "ACTIV METIERS",
   "ACTIV'CONSULT'TRAINING",
   "ACTUAL GROUP",
@@ -436,7 +435,6 @@ export const cfaCompanyList = [
   "CFA MARSEILLE",
   "CFA PETITE CAMARGUE",
   "CFA PHARMACIE DE L'AFPPREC - ASSOCIATION POUR LA FORMATION DES PREPARATEURS EN PHARMACIE DE LA REGION CVDL",
-  "CFA",
   "CFAI ALSACE",
   "CFAI D AUVERGNE",
   "CFAI DE BRETAGNE",
@@ -541,7 +539,6 @@ export const cfaCompanyList = [
   "CPO.A BTP",
   "CREAFI",
   "CREALISE",
-  "CREATIVE",
   "CROIX ROUGE FRANCAISE",
   "CRS PROF PERFECTIONNEMENT PHARMACIE",
   "CTI",
@@ -833,7 +830,6 @@ export const cfaCompanyList = [
   "FOOD'N CO LEARNING",
   "FOR YOU ONLY",
   "FORCE FORMAT COMPET EFFIC SAVOIR FAIRE",
-  "FORCES",
   "FORE ILES DU NORD",
   "FOREACH ACADEMY",
   "FOREZ INSERTION FORMATION CONSEIL",
@@ -966,7 +962,6 @@ export const cfaCompanyList = [
   "HAUTES ETUDES APPLIQUEES DU DROIT (HEAD)",
   "HBE",
   "HECTEN",
-  "HELP",
   "HENRIMAN FORMATION",
   "HETIC",
   "HEXAGONE",
@@ -1060,7 +1055,6 @@ export const cfaCompanyList = [
   "INHNI VILLENEUVE-D’ASCQ",
   "INHNI",
   "INITIATIVES FORMATION",
-  "INITIATIVES",
   "INNOVAFORMATION",
   "INNOVATION DEVELOPPEMENT FORMATION",
   "INNOV'EDUC",
@@ -1252,7 +1246,6 @@ export const cfaCompanyList = [
   "LA PLATEFORME FORMATION",
   "LA PRADETTE",
   "LA SOLIVE",
-  "LA SOURCE",
   "L'ACADEMIE BY MG",
   "L'ACADEMIE DES SERVICES AUX FAMILLES",
   "L'ACADEMIE OUI CARE",
@@ -1265,7 +1258,6 @@ export const cfaCompanyList = [
   "LE 101",
   "LE BIHAN ACADEMIE (LBA)",
   "LE CENTRE DES APPRENTIS D'ALSACE",
-  "LE CHIFFRE",
   "LE PORTRAIT",
   "LE STUDIO",
   "LEADER ACADEMY",
@@ -1616,7 +1608,6 @@ export const cfaCompanyList = [
   "RETRAVAILLER DANS L OUEST",
   "REUNI'FORM",
   "REUNION APPRENTISSAGE FORMATION",
-  "REUSSIR",
   "RHREFLEX",
   "RI7",
   "RIG UP FORMATION",
@@ -1792,10 +1783,8 @@ export const cfaCompanyList = [
   "UPYZI",
   "VAELIA",
   "VAL BUSINESS SCHOOL",
-  "VALEUR AJOUTEE",
   "VIA FORMATION",
   "VIATICUS PARDASA",
-  "VICTORIA",
   "VOL GROUPE CONSEILS, FORMATIONS",
   "VOLTAIRE BUSINESS SCHOOL",
   "VOLTEE",
